+++ tests/files/deep/b.js
@@ -1,2 +1,3 @@
 import * as c from './c'
 export { c }
+export default 'b'